Tidying methods for gamlss objects

Tidying methods for "gamlss" objects from the gamlss package.

# S3 method for gamlss
tidy(x, quick = FALSE, conf.int = FALSE, conf.level = 0.95, ...)

Arguments

x

A "gamlss" object

quick

Whether to perform a fast version, and return only the coefficients

conf.int

whether to return confidence intervals

conf.level

confidence level for CI

...

arguments passed to confint.gamlss

Value

All tidying methods return a data.frame without rownames, whose structure depends on the method chosen. A tibble with one row for each coefficient, containing columns:

parameter

type of coefficient being estimated: mu, sigma, nu, or tau

term

term in the model being estimated and tested

estimate

estimated coefficient

std.error

standard error

statistic

t-statistic

p.value

two-sided p-value

Examples

if (requireNamespace("gamlss", quietly = TRUE) &&
    requireNamespace("gamlss.data", quietly = TRUE)) {
    data(abdom, package="gamlss.data")
    if (FALSE) {
         mod <- gamlss(y~pb(x), sigma.fo=~pb(x), family=BCT,
                       data=abdom, method=mixed(1,20))
    }
    ## load stored object
    mod <- readRDS(system.file("extdata", "gamlss_example.rds",
                   package="broom.mixed"))
    tidy(mod)
}